Upgraded SLF4J to version 1.3.0
Fix for bug qfj-137 - modified the implementation of SLF4JLog to recognize the LocationAwareLogger and pass
the appropriate caller fully-qualified class name so that the right calling class is displayed.

This meant modifications to the LoggerFactory interface to introduce another parameter to the create() method - the callerFQCN. 
Each specfic LoggerFactory implementation specifies the right class name when the logger is created.
This was necessary so that CompositeLogger would work as well - otherwise, SLF4J logger wouldn't work if it was a nested logger inisde a CompositeLogger



git-svn-id: http://svn.code.sf.net/p/quickfixj/code/trunk@612 40aee5e9-8910-4a9f-ab63-41bea6217522